BIG CHANGES! task scheduler rework, font and image loader, base64, type registery, image explorer, table explorer, function explorer, instance getValue and setValue rework, tons more!

+++ include/classes/vector2.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
 #include "lua.h"
-#include <raylib.h>
+#include "raylib.h"
 
 namespace fakeroblox {
 
-int pushVector2(lua_State* L, double x, double y);
+int pushVector2(lua_State* L, float x, float y);
 int pushVector2(lua_State* L, Vector2 vector2);
 Vector2* lua_checkvector2(lua_State* L, int narg);
